remove global variables, moar docstrings
