+++ infra_event_notifier/backends/jira.py
@@ -1,1 +1,11 @@
-# FIXME
+from typing import Mapping
+
+
+def notify_jira(
+    title: str,
+    text: str,
+    tags: Mapping[str, str],
+    jira_api_key: str,
+) -> None:
+    # TODO: implement
+    pass
