+++ infra_event_notifier/backends/slack.py
@@ -1,1 +1,7 @@
-# FIXME
+def notify_slack(
+    title: str,
+    text: str,
+    slack_api_key: str,
+) -> None:
+    # TODO: implement
+    pass
